Reorganise directory structure to follow Ruby convention. Add skeleton command-line app.
