+++ bin/meetup_thingy
@@ -1,0 +1,7 @@
+#! /usr/bin/env ruby
+
+$LOAD_PATH.unshift("#{File.dirname(__FILE__)}/../lib")
+
+require 'meetup_thingy'
+
+MeetupThingy::App.start
