+++ lib/meetup_thingy.rb
@@ -1,0 +1,5 @@
+require 'meetup_thingy/app'
+require 'meetup_thingy/event_finder'
+require 'meetup_thingy/event_list_file_writer'
+require 'meetup_thingy/meetup_api'
+require 'meetup_thingy/meetup_file'
