+++ lib/meetup_thingy/app.rb
@@ -1,0 +1,7 @@
+require 'thor'
+
+module MeetupThingy
+  class App < Thor
+
+  end
+end
